+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## unreleased
+
+### Features
+
+* Add support for groups. These can provide default settings to checks,
+  and rate limiting for alerts. Checks can be members of any number of groups,
+  and alerts will be suppressed if any goes over the rate limit.
+
 ## 0.8.0 - 2025-06-13
 
 ### Changes